projectBrightMoney/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
black==22.10.0
celery==5.2.7
certifi==2022.9.24
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2
django-cors-headers==3.13.0
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
kombu==5.2.4
mypy-extensions==0.4.3
nulltype==2.3.1
packaging==21.3
pathspec==0.10.2
phonenumbers==8.13.0
plaid-python==11.2.0
platformdirs==2.5.4
prompt-toolkit==3.0.33
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
redis==4.3.5
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
